refactor: separate the responsibility of filters from the albums page
